output
Enter the No. of Rows: 3
Enter the No. of Columns: 4
Enter clean status for each cell (1 - dirty, 0 - clean)
1 0 0 0
0 1 0 1
1 0 1 1

The Floor matrix is as below:
 >1<   0    0    0  
  0    1    0    1  
  1    0    1    1  

The Floor matrix is as below:
 >0<   0    0    0  
  0    1    0    1  
  1    0    1    1  

The Floor matrix is as below:
  0   >0<   0    0  
  0    1    0    1  
  1    0    1    1  

The Floor matrix is as below:
  0    0   >0<   0  
  0    1    0    1  
  1    0    1    1  

The Floor matrix is as below:
  0    0    0   >0< 
  0    1    0    1  
  1    0    1    1  

The Floor matrix is as below:
  0    0    0    0  
  0    1    0   >1< 
  1    0    1    1  

The Floor matrix is as below:
  0    0    0    0  
  0    1    0   >0< 
  1    0    1    1  

The Floor matrix is as below:
  0    0    0    0  
  0    1   >0<   0  
  1    0    1    1  

The Floor matrix is as below:
  0    0    0    0  
  0   >1<   0    0  
  1    0    1    1  

The Floor matrix is as below:
  0    0    0    0  
  0   >0<   0    0  
  1    0    1    1  

The Floor matrix is as below:
  0    0    0    0  
 >0<   0    0    0  
  1    0    1    1  

The Floor matrix is as below:
  0    0    0    0  
  0    0    0    0  
 >1<   0    1    1  

The Floor matrix is as below:
  0    0    0    0  
  0    0    0    0  
 >0<   0    1    1  

The Floor matrix is as below:
  0    0    0    0  
  0    0    0    0  
  0   >0<   1    1  

The Floor matrix is as below:
  0    0    0    0  
  0    0    0    0  
  0    0   >1<   1  

The Floor matrix is as below:
  0    0    0    0  
  0    0    0    0  
  0    0   >0<   1  

The Floor matrix is as below:
  0    0    0    0  
  0    0    0    0  
  0    0    0   >1< 

The Floor matrix is as below:
  0    0    0    0  
  0    0    0    0  
  0    0    0   >0<